docker-requirements.txt
accelerate==0.24.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
auto-gptq==0.5.1
bitsandbytes==0.41.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.1
datasets==2.14.6
dill==0.3.7
diskcache==5.6.3
distro==1.8.0
einops==0.7.0
fastapi==0.104.0
filelock==3.13.0
FLAML==2.1.1
frozenlist==1.4.0
fsspec==2023.10.0
gekko==1.0.6
greenlet==3.0.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.325
langsmith==0.0.53
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.1
openai==1.3.7
optimum==1.14.1
packaging==23.2
pandas==2.1.2
peft==0.6.2
Pillow==9.3.0
protobuf==4.25.1
psutil==5.9.6
pyarrow==13.0.0
pyautogen==0.2.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rouge==1.0.1
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentencepiece==0.1.99
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.66.1
transformers==4.34.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
xxhash==3.4.1
yarl==1.9.2
